refactor(sdk): extract waitForReceiptResilient to _rpc-resilience.ts + dedup tx-submit tests
- Move waitForReceiptResilient out of uploader.ts into a shared
  src/_rpc-resilience.ts (internal, not re-exported), grouping it with the
  other shared-RPC hardening helper (safeWriteContract) and matching the
  existing __integration__/_rpc-resilience.ts test mirror. Drops the now-unused
  TransactionReceiptNotFoundError / WaitForTransactionReceiptTimeoutError / Hash
  imports from uploader.ts; behavior unchanged.
- _tx-submit.test.ts: collapse the repeated writeContract-shape boilerplate into
  a callPing(wallet, publicClient?, overrides?) helper.

205 passed | 23 skipped.
